enable ooc methods from magma
