+++ Addons/tsgLoadUnstructuredPoints.hpp
@@ -210,7 +210,7 @@ inline void loadUnstructuredDataL2tmpl(double const data_points[], int num_data,
     Data2D<scalar_type> coefficients =
         [&]()->Data2D<scalar_type>{
             #ifdef Tasmanian_ENABLE_HIP
-            return generateCoefficients<scalar_type>(data_points, num_data, model_values, tolerance, grid);
+            if (grid.isFourier()) return generateCoefficients<scalar_type>(data_points, num_data, model_values, tolerance, grid);
             #endif
             if (acceleration->mode == accel_gpu_cuda and hasGPUBasis(grid)){
                 acceleration->setDevice();
